requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
discord.py==2.2.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pony==0.7.16
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
tabulate==0.9.0
urllib3==2.0.2
yarl==1.9.2
